fix bool types in commands

+++ klippy/extras/analog_probe.py
@@ -90,8 +90,8 @@ class AnalogProbe:
                                              " auto_th=%c auto_std_mul=%u"
                                              " tare_buf_len=%u cur_buf_len=%u"
                                              % (self.mcu_endstop._oid, self.mcu_endstop._pin, self.mcu_endstop._pullup,
-                                                self.trigger_sup, self.trigger_inf, int(self.threshold*10),
-                                                self.auto_threshold, int(self.auto_std_multiplier*100),
+                                                int(self.trigger_sup), int(self.trigger_inf), int(self.threshold*10),
+                                                int(self.auto_threshold), int(self.auto_std_multiplier*100),
                                                 self.tare_buffer_len, self.current_buffer_len))
         self.mcu_endstop._mcu.add_config_cmd(
             "analog_probe_home oid=%d clock=0 sample_ticks=0 sample_count=0"
@@ -170,7 +170,7 @@ class AnalogProbe:
             self.auto_std_multiplier = gcmd.get_float("STD_MULTIPLIER", 5.0)
         else:
             self.threshold = gcmd.get_float("THRESHOLD", 0.5)
-        self.mcu_endstop._set_threshold_cmd.send([self.mcu_endstop._oid, int(self.threshold*10), self.auto_threshold, int(self.auto_std_multiplier*100)])
+        self.mcu_endstop._set_threshold_cmd.send([self.mcu_endstop._oid, int(self.threshold*10), int(self.auto_threshold), int(self.auto_std_multiplier*100)])
 
     def cmd_PRINT_CURRENT_VALUES(self, gcmd):
         gcmd.respond_info("Raw: %.1f, Current: %.1f, Tare: %.1f, Threshold: %.1f" %           \
